@@ -12,16 +12,16 @@
         }
     ],
     "require": {
-        "php": "^7.4||^8.0",
-        "endroid/data-sanitize": "^1.4.5",
+        "php": "^8.0",
+        "endroid/data-sanitize": "^1.6",
         "endroid/installer": "^1.2.2",
         "endroid/simple-spreadsheet": "^3.0",
-        "symfony/config": "^4.4||^5.0||^6.0",
-        "symfony/dependency-injection": "^4.4||^5.0||^6.0",
-        "symfony/http-foundation": "^4.4||^5.0||^6.0",
-        "symfony/http-kernel": "^4.4||^5.0||^6.0",
-        "symfony/routing": "^4.4||^5.0||^6.0",
-        "twig/twig": "^2.0||^3.0"
+        "symfony/config": "^5.4||^6.0",
+        "symfony/dependency-injection": "^5.4||^6.0",
+        "symfony/http-foundation": "^5.4||^6.0",
+        "symfony/http-kernel": "^5.4||^6.0",
+        "symfony/routing": "^5.4||^6.0",
+        "twig/twig": "^3.0"
     },
     "require-dev": {
         "endroid/quality": "dev-master"
@@ -45,6 +45,9 @@
         }
     },
     "config": {
-        "sort-packages": true
+        "sort-packages": true,
+        "allow-plugins": {
+            "endroid/installer": true
+        }
     }
 }
