@@ -2,13 +2,6 @@
 
 declare(strict_types=1);
 
-/*
- * (c) Jeroen van den Enden <info@endroid.nl>
- *
- * This source file is subject to the MIT license that is bundled
- * with this source code in the file LICENSE.
- */
-
 namespace Endroid\DataSanitizeBundle\Controller;
 
 use Endroid\DataSanitizeBundle\Configuration;
@@ -18,20 +11,11 @@ use Twig\Environment;
 
 final class MenuController
 {
-    /** @var Environment */
-    private $templating;
-
-    /** @var RouterInterface */
-    private $router;
-
-    /** @var Configuration */
-    private $configuration;
-
-    public function __construct(Environment $templating, RouterInterface $router, Configuration $configuration)
-    {
-        $this->templating = $templating;
-        $this->router = $router;
-        $this->configuration = $configuration;
+    public function __construct(
+        private Environment $templating,
+        private RouterInterface $router,
+        private Configuration $configuration
+    ) {
     }
 
     public function __invoke(): Response
