@@ -2,13 +2,6 @@
 
 declare(strict_types=1);
 
-/*
- * (c) Jeroen van den Enden <info@endroid.nl>
- *
- * This source file is subject to the MIT license that is bundled
- * with this source code in the file LICENSE.
- */
-
 namespace Endroid\DataSanitizeBundle\Controller;
 
 use Endroid\DataSanitize\SanitizerFactory;
@@ -19,17 +12,15 @@ use Symfony\Component\HttpFoundation\Response;
 
 final class MergeController
 {
-    private $configuration;
-    private $sanitizerFactory;
-
-    public function __construct(Configuration $configuration, SanitizerFactory $sanitizerFactory)
-    {
-        $this->configuration = $configuration;
-        $this->sanitizerFactory = $sanitizerFactory;
+    public function __construct(
+        private Configuration $configuration,
+        private SanitizerFactory $sanitizerFactory
+    ) {
     }
 
     public function __invoke(Request $request, string $name): Response
     {
+        /** @var array<string> $sourceIds */
         $sourceIds = (array) $request->query->get('sources');
         if (0 == count($sourceIds)) {
             return new JsonResponse([
