@@ -2,13 +2,6 @@
 
 declare(strict_types=1);
 
-/*
- * (c) Jeroen van den Enden <info@endroid.nl>
- *
- * This source file is subject to the MIT license that is bundled
- * with this source code in the file LICENSE.
- */
-
 namespace Endroid\DataSanitizeBundle\Controller;
 
 use Endroid\DataSanitize\SanitizerFactory;
@@ -18,13 +11,10 @@ use Symfony\Component\HttpFoundation\Response;
 
 final class StateController
 {
-    private $configuration;
-    private $sanitizerFactory;
-
-    public function __construct(Configuration $configuration, SanitizerFactory $sanitizerFactory)
-    {
-        $this->configuration = $configuration;
-        $this->sanitizerFactory = $sanitizerFactory;
+    public function __construct(
+        private Configuration $configuration,
+        private SanitizerFactory $sanitizerFactory
+    ) {
     }
 
     public function __invoke(string $name): Response
